Revert TerminalSize type duplication, use shared types
PR #133367 duplicated TerminalSize and TerminalSizeQueue types between
kubectl/pkg/util/term and client-go/tools/remotecommand to avoid heavy
dependencies. However, these types themselves have zero dependencies - only
other files in the remotecommand package use websocket/spdystream.

This duplication broke downstream consumers that use both packages (e.g.,
cri-tools), requiring workarounds:
https://github.com/kubernetes-sigs/cri-tools/pull/1955/files#diff-c72419df07499621d9866069c71e99656016c3149f71040424b8bd1af4e7ec07L430-R452

This change moves the types to k8s.io/apimachinery/pkg/util/remotecommand
(which already contains remotecommand protocol constants) and updates both
kubectl and client-go to use type aliases. This eliminates duplication while
achieving the original goal of package decoupling.
